import bodov z prijimacok, filtrovanie podla bodov z prijimacok

+++ bakalarka-backend/database/migrations/1544038708090_ais_admissions_schema.js
@@ -118,10 +118,10 @@ class AisAdmissionsSchema extends Schema {
       table.string("Zvol_predmet")
       table.string("Zvol_predmet_1")
 
-      table.integer("Externá_maturita_z_cudzieho_jazyka_ECJ")
-      table.integer("Externá_maturita_z_matematiky_EM")
-      table.integer("Písomný_test_z_matematiky_SCIO_PTM")
-      table.integer("Všeobecné_študijné_predpoklady_SCIO_VŠP")
+      table.decimal("Externá_maturita_z_cudzieho_jazyka_ECJ")
+      table.decimal("Externá_maturita_z_matematiky_EM")
+      table.decimal("Písomný_test_z_matematiky_SCIO_PTM")
+      table.decimal("Všeobecné_študijné_predpoklady_SCIO_VŠP")
     })
   }
